create form div is removed after a new event is created
