+++ scripts/eventFactory.js
@@ -1,18 +1,13 @@
 //Kristen
-//purpose create an event object
+//function for creating an event object
 
-const eventId = require("./idGenerator")
-const setLocalStorage = require("./setLocalStorage")
 const getLocalStorage = require("./getLocalStorage")
 const getSessionStorage = require("./getSessionStorage")
 const addEventList = require("./eventListController")
 
 const storedDb = getLocalStorage()
 
-let eventFactory = function(eventName, date, location, time, description, eventId) {
-    let lastEventId = storedDb.events[storedDb.events.length- 1] ||  {eventId: 0}
-    let eventIdFactory = eventId(lastEventId.eventId)
-    
+let eventFactory = function(eventId, eventName, date, location, time, description) {
 
     const newEvent = Object.create(null, {
         "userId": {
@@ -20,7 +15,7 @@ let eventFactory = function(eventName, date, location, time, description, eventI
             "enumerable": true
         },
         "eventId": {
-            "value": (eventId === null) ? eventIdFactory.next().value : eventId,
+            "value": eventId,
             "enumerable": true
         },
         "eventName": {
@@ -54,22 +49,4 @@ let eventFactory = function(eventName, date, location, time, description, eventI
     addEventList(newEvent)
 }
 
-const createEvent = () => {
-    let eName = document.getElementById("event_name").value
-    let eDate = document.getElementById("event_date").value
-    let eLocation = document.getElementById("event_location").value
-    let eTime = document.getElementById("event_time").value
-    let eDescription = document.getElementById("event_description").value
-
-    if (eName === "" || eDate === "" || eLocation === "" || eTime === "" || eDescription === "") {
-        alert("Please fill out all fields")
-    } else {
-        eventFactory(eName, eDate, eLocation, eTime, eDescription)
-
-        setLocalStorage(storedDb)
-
-    }
-    document.getElementById("event_formContent").innerHTML = " "
-}
-
-module.exports = createEvent
+module.exports = eventFactory
