+++ scripts/eventUpdateController.js
@@ -1,0 +1,2 @@
+//Kristen
+//functionality when the update button is pressed
